update: allow update absent nullable fields
Update operations can't insert with gaps. This patch changes the
behavior so that the update operation fills the missing fields
with zeros. For example we create space s = box.schema.create_space('s'),
then create index for this space pk = s:create_index('pk'), and then
insert tuple in space s:insert{1, 2}. After all of this we try to update
this tuple s:update({1}, {{'!', 5, 6}}), in previous version this operation
fails with ER_NO_SUCH_FIELD_NO error, and after this patch it's finished
with success and there is [1, 2, null, null, 6] tuple in space.
Closes #3378

@TarantoolBot document
Title: Allow update absent nullable fields
Update operations can't insert with gaps. This patch changes the
behavior so that the update operation fills the missing fields
with zeros. For example we create space s = box.schema.create_space('s'),
then create index for this space pk = s:create_index('pk'), and then
insert tuple in space s:insert{1, 2}. After all of this we try to update
this tuple s:update({1}, {{'!', 5, 6}}), in previous version this operation
fails with ER_NO_SUCH_FIELD_NO error, and after this patch it's finished
with success and there is [1, 2, null, null, 6] tuple in space.